Add isSaved endpoint
