+++ src/services/UserService.ts
@@ -1,6 +1,4 @@
 import { NotFoundError } from 'routing-controllers';
-import { PostAndUserUuidParam } from 'src/api/validators/GenericRequests';
-import { PostModel } from 'src/models/PostModel';
 import { Service } from 'typedi';
 import { EntityManager } from 'typeorm';
 import { InjectManager } from 'typeorm-typedi-extensions';
@@ -50,32 +48,6 @@ export class UserService {
       const user = await postRepository.getUserByPostId(params.id);
       if (!user) throw new NotFoundError('Post not found!');
       return user;
-    });
-  }
-
-  public async savePost(params: PostAndUserUuidParam): Promise<PostModel> {
-    return this.transactions.readOnly(async (transactionalEntityManager) => {
-      const userRepository = Repositories.user(transactionalEntityManager);
-      const user = await userRepository.getUserById(params.userId);
-      if (!user) throw new NotFoundError('User not found!');
-      const postRepository = Repositories.post(transactionalEntityManager);
-      const post = await postRepository.getPostById(params.postId);
-      if (!post) throw new NotFoundError('Post not found!');
-      await userRepository.savePost(user, post);
-      return post
-    });
-  }
-
-  public async unsavePost(params: PostAndUserUuidParam): Promise<PostModel> {
-    return this.transactions.readOnly(async (transactionalEntityManager) => {
-      const userRepository = Repositories.user(transactionalEntityManager);
-      const user = await userRepository.getUserById(params.userId);
-      if (!user) throw new NotFoundError('User not found!');
-      const postRepository = Repositories.post(transactionalEntityManager);
-      const post = await postRepository.getPostById(params.postId);
-      if (!post) throw new NotFoundError('Post not found!');
-      await userRepository.unsavePost(user, post);
-      return post
     });
   }
